feat(redliner): Create components prop
