+++ src/components/RedLiner/helpers/components.ts
@@ -1,0 +1,15 @@
+import { ComponentType } from 'react';
+import InfoBox, { IInfoBoxProps } from '../../InfoBox/InfoBox';
+
+export type ComponentsConfig = {
+  InfoBox: ComponentType<IInfoBoxProps>;
+};
+
+const defaultComponents = {
+  InfoBox,
+};
+
+export const getCurrentComponents = (providedComponents: ComponentsConfig) => ({
+  ...defaultComponents,
+  ...providedComponents,
+});
